WIP: Refactor out PentestGame, PentestDivision, and PointGame
This is very much unfinished however this removes many of the STI models that were added, moving to a much simpler setup.

This also changes our category relation to be a many-to-many relationship, and adds the beginnings of support for different board designs. I have currently added code for Jeopardy style, and have added placeholders for "Teams X Challenges" style for embedded CTF and "Multiple Categories" for @noraj, however the implementations for the last two are unfinished.