WIP: Refactor application slightly (again) to logically structure challenges
Previously I was attempting to structure things based on their functionality. That was not working well due to the rigidity of rails_admin. Items are now more logically grouped, with logic broken out into the lib/ directory for things such as share challenge calculation. This means defensive challenges can be controlled by the same logic that controls ShareChallenges, just without the code being duplicated and without the two models being at all dependent on one another. This frees up the module dependency structure to better reflect the logical structure in the app. For example, previously I was attempting to create a superclass for ShareChallenges and PentestChallenges. This doesn't work since the calculations for PentestChallenges are actually performed on a per-flag basis, since the points are calculated at the per-flag level. Now there is a shared module with the calculation logic for both ShareChallenges and DefenseFlags, and ShareChallenges and PentestChallenges only share a single base class, Challenge.